Refactored testserver into multiple files
